all search functions can now com, triager can flag invalidment, now when closing remarks will be added to bug
along with other bug fixes here and there

+++ src/com/company/TriagerViewAllBugsController.java
@@ -19,7 +19,7 @@ class TriagerViewAllBugsController
     public TriagerViewAllBugsController(String bugFile)
     {
         this.bugFile = bugFile;
-        flagDup();
+        //flagDup();
     }
 
     public HashMap<Integer, String> getFilesMap()
@@ -32,7 +32,69 @@ class TriagerViewAllBugsController
         BugFileAccess bd = new BugFileAccess();
         return bd.getTitleMap();
     }
-    public void flagDup()
+
+    public String flagInv()
+    {
+        BugFileAccess bd = new BugFileAccess();
+        HashMap<String, Bug> bugMap = bd.getBugMap();
+        SearchController sc = new SearchController(bugFile);
+        bugFile = sc.searchByTitle();
+        Bug b = new Bug();
+
+        if(bugFile.contains("!"))
+        {
+
+            String fullFile = "Bugs/" + bugFile;
+            for(Map.Entry me : bugMap.entrySet())
+            {
+                if(me.getKey().equals(fullFile))
+                {
+                    b = (Bug) me.getValue();
+                    b.setSolved("closed (FLAGGED AS INVALID)");
+                    b.setAssignDeveloper("N/A due to invalid bug");
+                    String fileID = b.getID();
+                    File f = new File(bugFile);
+                    f.delete();
+                    b.writeBugToFile();
+    
+                }
+            }
+            
+            //get rid of ! in bugfiledatabase.txt
+            CommentAccess cd = new CommentAccess();
+            bd.setFileName(bugFile);
+            cd.setFileName(bugFile);
+    
+            //delete the old file with the "!"
+            File toDel = new File("Bugs/" + bugFile);
+            toDel.delete();
+            return "true";
+        }
+        else
+        {
+            String s = "";
+            String fullFile = "Bugs/" + bugFile;
+            for(Map.Entry me : bugMap.entrySet())
+            {
+                if(me.getKey().equals(fullFile))
+                {
+                    b = (Bug) me.getValue();
+                    s = b.getAssignDeveloper();
+                }
+            }
+
+            if(s.isEmpty())
+            {
+                return "false";
+            }
+            else
+            {
+                return "dev";
+            }
+        }
+    }
+
+    public String flagDup()
     {
         //get the bugfile
         //match to the corresponding bug in the bugmap 
@@ -43,30 +105,56 @@ class TriagerViewAllBugsController
         bugFile = sc.searchByTitle();
         Bug b = new Bug();
 
-        String fullFile = "Bugs/" + bugFile;
-        for(Map.Entry me : bugMap.entrySet())
+        if(bugFile.contains("!"))
         {
-            if(me.getKey().equals(fullFile))
-            {
-                b = (Bug) me.getValue();
-                b.setSolved("closed (FLAGGED AS DUPLICATE)");
-                b.setAssignDeveloper("N/A due to duplicate bug");
-                String fileID = b.getID();
-                File f = new File(bugFile);
-                f.delete();
-                b.writeBugToFile();
 
+            String fullFile = "Bugs/" + bugFile;
+            for(Map.Entry me : bugMap.entrySet())
+            {
+                if(me.getKey().equals(fullFile))
+                {
+                    b = (Bug) me.getValue();
+                    b.setSolved("closed (FLAGGED AS DUPLICATE)");
+                    b.setAssignDeveloper("N/A due to duplicate bug");
+                    String fileID = b.getID();
+                    File f = new File(bugFile);
+                    f.delete();
+                    b.writeBugToFile();
+    
+                }
+            }
+            
+            //get rid of ! in bugfiledatabase.txt
+            CommentAccess cd = new CommentAccess();
+            bd.setFileName(bugFile);
+            cd.setFileName(bugFile);
+    
+            //delete the old file with the "!"
+            File toDel = new File("Bugs/" + bugFile);
+            toDel.delete();
+            return "true";
+        }
+        else
+        {
+            String s = "";
+            String fullFile = "Bugs/" + bugFile;
+            for(Map.Entry me : bugMap.entrySet())
+            {
+                if(me.getKey().equals(fullFile))
+                {
+                    b = (Bug) me.getValue();
+                    s = b.getAssignDeveloper();
+                }
+            }
+
+            if(s.isEmpty())
+            {
+                return "false";
+            }
+            else
+            {
+                return "dev";
             }
         }
-        
-        //get rid of ! in bugfiledatabase.txt
-        CommentAccess cd = new CommentAccess();
-        bd.setFileName(bugFile);
-        cd.setFileName(bugFile);
-
-        //delete the old file with the "!"
-        File toDel = new File("Bugs/" + bugFile);
-        toDel.delete();
-        
     }
 }
